Added enrichment analysis preview
Functional enrichment analysis is executed at the end of the analysis rom ProteoSign. The results are saved per pair of conditions in text tab seperated files and are zipped to be returned to the user. Added a preview of these files, the user can choose which pair of conditions' analysis to examine and double clicking on a term they are redirected to a site with more information. This feature uses Tabulator, an MIT-licensed javascrip library for interactive tables