Added Venn Diagrams generation
Venn diagrams are a very useful source of information regarding the reproducibility of the experiment, Using the VennDiagram pacjage in R we added the functionality of producing Venn diagrams of quantified proteins across conditions, biological replicates and technical replicates of each biological replicate. The result is shown back to the user and saved to the results zip file,. The code has also been copied to Plot Generator.R so that the user might tweak it and alter its appearance.