Split perform_analysis to functions
peform analysis was the main function of MSdiffexp.R  that prepared data to be fed to main procedure functions. The data loading and preparation was unorganized and uncommented, This commit adds new functions that do parts of this loading and preparation and hides unnecessary code from perform analysis and adds many lines of comments with examples to make the code easier to read.